Rename ThreadSampler native code files to thread_sampler
