+++ tracer/src/Datadog.Trace.ClrProfiler.Native/CMakeLists.txt
@@ -230,7 +230,7 @@ add_library("SignalFx.Tracing.ClrProfiler.Native.static" STATIC
         util.cpp
         calltarget_tokens.cpp
         rejit_handler.cpp
-        ThreadSampler.cpp
+        thread_sampler.cpp
         lib/coreclr/src/pal/prebuilt/idl/corprof_i.cpp
         ${GENERATED_OBJ_FILES}
 )
