+++ tracer/src/Datadog.Trace.ClrProfiler.Native/Datadog.Trace.ClrProfiler.Native.vcxproj
@@ -206,7 +206,7 @@
     <ClInclude Include="rejit_handler.h" />
     <ClInclude Include="stats.h" />
     <ClInclude Include="string.h" />
-    <ClInclude Include="ThreadSampler.h" />
+    <ClInclude Include="thread_sampler.h" />
     <ClInclude Include="util.h" />
     <ClInclude Include="version.h" />
   </ItemGroup>
@@ -224,7 +224,7 @@
     <ClCompile Include="miniutf.cpp" />
     <ClCompile Include="rejit_handler.cpp" />
     <ClCompile Include="string.cpp" />
-    <ClCompile Include="ThreadSampler.cpp" />
+    <ClCompile Include="thread_sampler.cpp" />
     <ClCompile Include="util.cpp" />
   </ItemGroup>
   <ItemGroup>
