+++ tracer/src/Datadog.Trace.ClrProfiler.Native/thread_sampler.h
